+++ all-is-cubes/src/math/rotation.rs
@@ -216,7 +216,9 @@ impl GridRotation {
         }
     }
 
-    // TODO: public? do we want this to be our API? should this also be a From impl?
+    /// Returns the basis vectors for the unrotated coordinate system in
+    /// the rotated coordinate system.
+    // TODO: public? do we want this to be our API?
     #[doc(hidden)]
     #[inline]
     #[rustfmt::skip] // dense data layout
@@ -321,51 +323,6 @@ impl GridRotation {
         Gridgid {
             rotation: self,
             translation: offset(basis.x, size) + offset(basis.y, size) + offset(basis.z, size),
-        }
-    }
-
-    /// Expresses this rotation as a matrix which rotates “in place” the
-    /// points within the volume defined by coordinates in the range [0, size].
-    ///
-    /// That is, a [`GridAab`] of that volume will be unchanged by rotation:
-    ///
-    /// ```
-    /// use all_is_cubes::block::Resolution;
-    /// use all_is_cubes::math::{GridAab, GridRotation};
-    ///
-    /// let b = GridAab::for_block(Resolution::R8);
-    /// let rotation = GridRotation::CLOCKWISE.to_positive_octant_matrix(8);
-    /// assert_eq!(b.transform(rotation), Some(b));
-    /// ```
-    ///
-    /// Such matrices are suitable for rotating the voxels of a block, provided
-    /// that the coordinates are then transformed with [`GridMatrix::transform_cube`],
-    /// *not* [`GridMatrix::transform_point`](cgmath::Transform::transform_point)
-    /// (due to the lower-corner format of cube coordinates).
-    /// ```
-    /// # use all_is_cubes::math::{GridAab, GridPoint, GridRotation};
-    /// let rotation = GridRotation::CLOCKWISE.to_positive_octant_matrix(4);
-    /// assert_eq!(rotation.transform_cube(GridPoint::new(0, 0, 0)), GridPoint::new(3, 0, 0));
-    /// assert_eq!(rotation.transform_cube(GridPoint::new(3, 0, 0)), GridPoint::new(3, 0, 3));
-    /// assert_eq!(rotation.transform_cube(GridPoint::new(3, 0, 3)), GridPoint::new(0, 0, 3));
-    /// assert_eq!(rotation.transform_cube(GridPoint::new(0, 0, 3)), GridPoint::new(0, 0, 0));
-    /// ```
-    ///
-    // TODO: add tests
-    pub fn to_positive_octant_matrix(self, size: GridCoordinate) -> GridMatrix {
-        fn offset(face: Face6, size: GridCoordinate) -> GridVector {
-            if face.is_positive() {
-                GridVector::zero()
-            } else {
-                face.normal_vector() * -size
-            }
-        }
-        let basis = self.to_basis();
-        GridMatrix {
-            x: basis.x.normal_vector(),
-            y: basis.y.normal_vector(),
-            z: basis.z.normal_vector(),
-            w: offset(basis.x, size) + offset(basis.y, size) + offset(basis.z, size),
         }
     }